Cleaning up the refactorization. Everything now resides in the src folder
